move ReportType back to checkers.h, keep functions in errorlogger.h
